+++ ICD.Connect.Telemetry.Crestron/FusionTelemetryBinding.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Globalization;
 using System.Linq;
 using ICD.Common.Utils.Attributes;
 using ICD.Common.Utils.Extensions;
@@ -25,7 +26,16 @@ namespace ICD.Connect.Telemetry.Crestron
 		public IFusionRoom FusionRoom { get; private set; }
 		public IFusionAsset Asset { get; private set; }
 
-		private FusionTelemetryBinding(IFusionRoom fusionRoom, ITelemetryItem getTelemetry, ITelemetryItem setTelemetry, IFusionAsset asset, FusionSigMapping mapping)
+		/// <summary>
+		/// Constructor.
+		/// </summary>
+		/// <param name="fusionRoom"></param>
+		/// <param name="getTelemetry"></param>
+		/// <param name="setTelemetry"></param>
+		/// <param name="asset"></param>
+		/// <param name="mapping"></param>
+		private FusionTelemetryBinding(IFusionRoom fusionRoom, ITelemetryItem getTelemetry, ITelemetryItem setTelemetry,
+		                               IFusionAsset asset, FusionSigMapping mapping)
 		{
 			FusionRoom = fusionRoom;
 			GetTelemetry = getTelemetry as IFeedbackTelemetryItem;
@@ -33,7 +43,7 @@ namespace ICD.Connect.Telemetry.Crestron
 			Asset = asset;
 
 			IUpdatableTelemetryNodeItem updatable = GetTelemetry as IUpdatableTelemetryNodeItem;
-			if(updatable != null)
+			if (updatable != null)
 				updatable.OnValueChanged += UpdatableOnValueChanged;
 
 			Mapping = mapping;
@@ -62,7 +72,7 @@ namespace ICD.Connect.Telemetry.Crestron
 		private void UpdateDigitalTelemetry()
 		{
 			FusionSigMapping singleMapping = Mapping;
-			if(singleMapping == null)
+			if (singleMapping == null)
 				return;
 
 			if (Mapping.Sig == 0)
@@ -70,15 +80,11 @@ namespace ICD.Connect.Telemetry.Crestron
 
 			bool digital = ((IFusionStaticAsset)Asset).ReadDigitalSig(singleMapping.Sig);
 
-			//Handle case where two seperate digitals are used to set true and false
+			// Handle case where two separate digitals are used to set true and false
 			if (SetTelemetry.ParameterCount == 0 && digital)
-			{
 				SetTelemetry.Invoke();
-			}
 			else
-			{
 				SetTelemetry.Invoke(digital);
-			}
 		}
 
 		private void UpdateAnalogTelemetry()
@@ -91,8 +97,28 @@ namespace ICD.Connect.Telemetry.Crestron
 				return;
 
 			ushort analog = ((IFusionStaticAsset)Asset).ReadAnalogSig(singleMapping.Sig);
-			object rescaledValue = GetRescaledAnalogValue(analog);
-			SetTelemetry.Invoke(rescaledValue);
+			object rescaledValue;
+
+			try
+			{
+				rescaledValue = GetRescaledAnalogValue(analog);
+			}
+			catch (Exception e)
+			{
+				ServiceProvider.GetService<ILoggerService>()
+				               .AddEntry(eSeverity.Error, "{0} - Failed to convert analog telemtry value - {1}", this, e.Message);
+				return;
+			}
+
+			try
+			{
+				SetTelemetry.Invoke(rescaledValue);
+			}
+			catch (Exception e)
+			{
+				ServiceProvider.GetService<ILoggerService>()
+				               .AddEntry(eSeverity.Error, "{0} - Failed to invoke SetTelemetry method - {1}", this, e.Message);
+			}
 		}
 
 		private void UpdateSerialTelemetry()
@@ -142,7 +168,7 @@ namespace ICD.Connect.Telemetry.Crestron
 
 		private void UpdateReservedSig()
 		{
-			if(Asset is IFusionStaticAsset)
+			if (Asset is IFusionStaticAsset)
 			{
 				if (Mapping.TelemetryGetName == DeviceTelemetryNames.ONLINE_STATE)
 				{
@@ -161,7 +187,7 @@ namespace ICD.Connect.Telemetry.Crestron
 			{
 				if (Mapping.TelemetryGetName == OccupancyTelemetryNames.OCCUPANCY_STATE)
 				{
-					bool value = ((eOccupancyState)GetTelemetry.Value) == eOccupancyState.Occupied;
+					bool value = (eOccupancyState)GetTelemetry.Value == eOccupancyState.Occupied;
 					((IFusionOccupancySensorAsset)Asset).SetRoomOccupied(value);
 					return;
 				}
@@ -184,14 +210,26 @@ namespace ICD.Connect.Telemetry.Crestron
 			if (singleMapping == null || GetTelemetry == null || GetTelemetry.Value == null)
 				return;
 
-			ushort analog = GetNewAnalogValue();
+			ushort analog;
+
+			try
+			{
+				analog = GetValueAsAnalog();
+			}
+			catch (Exception e)
+			{
+				ServiceProvider.GetService<ILoggerService>()
+				               .AddEntry(eSeverity.Error, "{0} - Failed to convert value to analog - {1}", this, e.Message);
+				return;
+			}
+
 			((IFusionStaticAsset)Asset).UpdateAnalogSig(singleMapping.Sig, analog);
 		}
 
 		private void UpdateSerialSig()
 		{
 			FusionSigMapping singleMapping = Mapping;
-			if (singleMapping == null || GetTelemetry == null ||GetTelemetry.Value == null)
+			if (singleMapping == null || GetTelemetry == null || GetTelemetry.Value == null)
 				return;
 
 			string serial = GetTelemetry.Value.ToString();
@@ -200,62 +238,48 @@ namespace ICD.Connect.Telemetry.Crestron
 
 		private object GetRescaledAnalogValue(ushort analog)
 		{
-			if(SetTelemetry == null || SetTelemetry.ParameterCount == 0)
+			if (SetTelemetry == null || SetTelemetry.ParameterCount == 0)
 				throw new InvalidOperationException("Set telemetry is null or has no parameters");
 
 			Type targetType = SetTelemetry.ParameterTypes[0];
 
-			if (!targetType.IsDecimalNumeric())
-			{
-				return Convert.ChangeType(analog, targetType, null);
-			}
+			RangeAttribute rangeAttribute =
+				GetTelemetry.PropertyInfo
+				            .GetCustomAttributes<RangeAttribute>()
+				            .FirstOrDefault();
 
-			RangeAttribute rangeAttribute = EnumerableExtensions.FirstOrDefault(GetTelemetry.PropertyInfo
-			                                                                                .GetCustomAttributes<RangeAttribute>(),
-			                                                                    null);
-			return rangeAttribute == null 
-				? Convert.ChangeType(analog, targetType, null) 
-				: rangeAttribute.RemapUshortToRange(analog);
+			if (rangeAttribute != null)
+				return Convert.ChangeType(rangeAttribute.RemapMinMax(analog), targetType, CultureInfo.InvariantCulture);
+
+			return RangeAttribute.Clamp(analog, targetType);
 		}
 
-		private ushort GetNewAnalogValue()
+		private ushort GetValueAsAnalog()
 		{
-			if (!GetTelemetry.ValueType.IsDecimalNumeric())
-			{
-				return (ushort)Convert.ChangeType(GetTelemetry.Value, typeof(ushort), null);
-			}
+			RangeAttribute rangeAttribute =
+				GetTelemetry.PropertyInfo
+				            .GetCustomAttributes<RangeAttribute>()
+				            .FirstOrDefault();
 
-			RangeAttribute rangeAttribute = EnumerableExtensions.FirstOrDefault(GetTelemetry.PropertyInfo
-			                                                                        .GetCustomAttributes<RangeAttribute>(), null);
-			if (rangeAttribute == null)
-			{
-				return (ushort)Convert.ChangeType(GetTelemetry.Value, typeof(ushort), null);
-			}
+			if (rangeAttribute != null)
+				return Convert.ToUInt16(rangeAttribute.ClampMinMaxThenRemap(GetTelemetry.Value, typeof(ushort)));
 
-			if (rangeAttribute.IsInRange(GetTelemetry.Value))
-			{
-				double value = (double)Convert.ChangeType(GetTelemetry.Value, typeof(double), null);
+			// If decimal map dec min-max to ushort min-max
+			Type propertyType = GetTelemetry.PropertyInfo.PropertyType;
+			if (propertyType.IsDecimalNumeric())
+				return Convert.ToUInt16(RangeAttribute.Remap(GetTelemetry.Value, typeof(ushort)));
 
-				//Cast value to double, since we only care about decimal types here 
-				//and all decimal types can be converted to double
-				return rangeAttribute.RemapRangeToUshort(value);
-			}
-
-			ServiceProvider.GetService<ILoggerService>()
-			               .AddEntry(eSeverity.Warning,
-			                         "Warning when converting analog telemtry value. Value {0} is outside expected range {1}-{2}. Value will be set to 0.",
-			                         GetTelemetry.Value,
-			                         rangeAttribute.Min,
-			                         rangeAttribute.Max);
-			return 0;
+			// Integral types are clamped
+			return Convert.ToUInt16(RangeAttribute.Clamp(GetTelemetry.Value, typeof(ushort)));
 		}
 
-		public static FusionTelemetryBinding Bind(IFusionRoom fusionRoom, ITelemetryProvider provider, FusionSigMapping mapping, uint assetId)
+		public static FusionTelemetryBinding Bind(IFusionRoom fusionRoom, ITelemetryProvider provider,
+		                                          FusionSigMapping mapping, uint assetId)
 		{
 			if (fusionRoom == null)
 				throw new ArgumentNullException("fusionRoom");
 
-			if(provider == null)
+			if (provider == null)
 				throw new ArgumentException("provider");
 
 			if (mapping == null)
@@ -268,7 +292,7 @@ namespace ICD.Connect.Telemetry.Crestron
 			ITelemetryItem getTelemetryItem = TelemetryService.GetTelemetryForProvider(provider, mapping.TelemetryGetName);
 			ITelemetryItem setTelemetryItem = TelemetryService.GetTelemetryForProvider(provider, mapping.TelemetrySetName);
 
-			if(getTelemetryItem == null && setTelemetryItem == null)
+			if (getTelemetryItem == null && setTelemetryItem == null)
 				throw new InvalidOperationException("Cannot create telemetry binding with neither a getter nor a setter.");
 
 			// If the sig number is 0, that indicates that the sig is special-handling